tests/ui: regenerate records
It seems counterintuitive that we need to regenerate ui fixtures because
of some work on passphrase.

The reason is that we have changed where the state (now session id) is
generated. We have moved it to Features, so it is generated already
after the Initialize message is received. It used to be generated later
- during the PassphraseReq/Ack workflow and in some tests we ask for
randomness between those steps.